All tests are working. project_id always the newest value (needs fixing)

+++ src/main/java/no/kristiania/database/AbstractDao.java
@@ -13,10 +13,10 @@ public abstract class AbstractDao<T> {
         this.dataSource = dataSource;
     }
 
-    protected T retrieve(long id, String sql) throws SQLException {
+    protected T retrieve(Integer id, String sql) throws SQLException {
         try (Connection connection = dataSource.getConnection()) {
             try (PreparedStatement statement = connection.prepareStatement(sql)) {
-                statement.setLong(1, id);
+                statement.setInt(1, id);
                 try (ResultSet rs = statement.executeQuery()) {
                     if (rs.next()) {
                         return mapRow(rs);
